+++ packages/css-tokenizer/src/reader.ts
@@ -24,7 +24,7 @@ export class Reader implements CodePointReader {
 		return this.cursor - 1;
 	}
 
-	advanceCodePoint(n = 1) {
+	advanceCodePoint(n = 1): void {
 		this.cursor += n;
 		this.representationEnd = this.cursor - 1;
 	}