+++ packages/media-query-list-parser/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Refactor `MediaFeatureBoolean` so that it follows the same structure as `MediaFeaturePlain` (breaking)
 - Change the `ParseError` interface, this is now a subclass of `Error` (breaking)
 - Add `getName` and `getNameToken` to all nodes that have a feature name.
+- Add `@custom-media` parsing.
 
 ### 1.0.0 (November 14, 2022)